Massively decreased the number of errors being logged to console.
The source of these errors is event listeners that can not be removed.
The only solution I could come up with was minimizing the amount of event listeners present.
A full rewrite of the movement code could get rid of the problem entirely but I have no intention of doing one at this time.

Signed-off-by: Michael Tarasov <71594701+MichaelTarasov21@users.noreply.github.com>